# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EndpointCreate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "base_url": {
            "default": "",
            "title": "Base Url",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelSpec"
            },
            "minItems": 1,
            "title": "Models",
            "type": "array"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "enum": [
              "anthropic",
              "openai_compat",
              "ollama"
            ],
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "name",
          "provider",
          "models"
        ],
        "title": "EndpointCreate",
        "type": "object"
      },
      "EndpointOut": {
        "properties": {
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "has_api_key": {
            "title": "Has Api Key",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelSpec"
            },
            "title": "Models",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "base_url",
          "has_api_key",
          "api_key_env",
          "models",
          "enabled"
        ],
        "title": "EndpointOut",
        "type": "object"
      },
      "EndpointUpdate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "models": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ModelSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Models"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "EndpointUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelSpec": {
        "properties": {
          "context_window": {
            "default": 8192,
            "minimum": 1024.0,
            "title": "Context Window",
            "type": "integer"
          },
          "cost_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost In"
          },
          "cost_out": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Out"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reasoning_class": {
            "default": "standard",
            "enum": [
              "micro",
              "light",
              "standard",
              "deep"
            ],
            "title": "Reasoning Class",
            "type": "string"
          },
          "structured": {
            "default": true,
            "title": "Structured",
            "type": "boolean"
          },
          "vision": {
            "default": false,
            "title": "Vision",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ],
        "title": "ModelSpec",
        "type": "object"
      },
      "OkResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PuttyU",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/model-endpoints": {
      "get": {
        "operationId": "list_endpoints_api_model_endpoints_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EndpointOut"
                  },
                  "title": "Response List Endpoints Api Model Endpoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Endpoints",
        "tags": [
          "providers"
        ]
      },
      "post": {
        "operationId": "create_endpoint_api_model_endpoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Endpoint",
        "tags": [
          "providers"
        ]
      }
    },
    "/api/model-endpoints/{endpoint_id}": {
      "delete": {
        "operationId": "delete_endpoint_api_model_endpoints__endpoint_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Endpoint",
        "tags": [
          "providers"
        ]
      },
      "put": {
        "operationId": "update_endpoint_api_model_endpoints__endpoint_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Endpoint",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}